+++ packages/components/rollup.config.js
@@ -5,7 +5,7 @@ import postcss from "rollup-plugin-postcss";
 import commonjs from "@rollup/plugin-commonjs";
 
 export default {
-  input: `src/*/index.ts`,
+  input: `src/*/index.{ts,tsx}`,
   plugins: [
     multiInput(),
     typescript({